+++ src/quorum/service/network.ts
@@ -1,6 +1,7 @@
 import { ethers } from 'ethers'
 import RLP from 'rlp'
-import { NetworkCreateType, NetworkGenerateType, GenesisJsonType, JoinNodeType, AddValidatorRemoteType, AddMemberRemoteType } from '../model/type/network.type'
+import { NetworkCreateType, NetworkGenerateType, JoinNodeType, AddValidatorRemoteType, AddMemberRemoteType } from '../model/type/network.type'
+import GenesisJsonYaml from '../model/yaml/network/gensisJson'
 import { AbstractService } from './Service.abstract'
 import ValidatorInstance from '../instance/validator'
 import ValidatorDockerComposeYaml from '../model/yaml/docker-compose/validatorDockerComposeYaml'
@@ -33,48 +34,12 @@ export default class Network extends AbstractService {
       alloc[`0x${x.account.replace(/^0x/, '').toLowerCase()}`] = { balance: x.amount }
     })
 
-    const genesisJson: GenesisJsonType = {
-      nonce: '0x0',
-      timestamp: `0x${Math.floor(Date.now() / 1000).toString(16)}`,
-      extraData,
-      gasLimit: '0xE0000000',
-      gasUsed: '0x0',
-      number: '0x0',
-      difficulty: '0x1',
-      coinbase: '0x0000000000000000000000000000000000000000',
-      mixHash: '0x63746963616c2062797a616e74696e65206661756c7420746f6c6572616e6365',
-      parentHash: '0x0000000000000000000000000000000000000000000000000000000000000000',
-      config: {
-        chainId: networkCreateConfig.chainId,
-        homesteadBlock: 0,
-        eip150Block: 0,
-        eip150Hash: '0x0000000000000000000000000000000000000000000000000000000000000000',
-        eip155Block: 0,
-        eip158Block: 0,
-        byzantiumBlock: 0,
-        constantinopleBlock: 0,
-        petersburgBlock: 0,
-        istanbulBlock: 0,
-        qbft: {
-          epochLength: 30000,
-          blockPeriodSeconds: 1,
-          emptyBlockPeriodSeconds: 3600,
-          requestTimeoutSeconds: 60,
-          policy: 0,
-          ceil2Nby3Block: 0,
-        },
-        transitions: [
-          {
-            block: 0,
-            transactionSizeLimit: 64,
-            contractSizeLimit: 64,
-          },
-        ],
-        txnSizeLimit: 64,
-        isQuorum: true,
-      },
-      alloc,
-    }
+    const genesisJsonYaml = new GenesisJsonYaml()
+    genesisJsonYaml.addExtraData(extraData)
+    genesisJsonYaml.addChainId(networkCreateConfig.chainId)
+    genesisJsonYaml.addAlloc(alloc)
+    const genesisJson = JSON.parse(genesisJsonYaml.getJsonString())
+
     this.bdkFile.createArtifactsFolder()
     this.bdkFile.createGenesisJson(genesisJson)
 
@@ -531,17 +496,6 @@ export default class Network extends AbstractService {
     this.bdkFile.createPrivateKey(dir, privateKey)
     this.bdkFile.createPublicKey(dir, publicKey)
     this.bdkFile.createAddress(dir, address)
-
-    return { privateKey, publicKey, address }
-  }
-
-  /** @ignore */
-  public createWalletAddress () {
-    // TODO: use Shawn's code generate key
-    const nodekey = ethers.Wallet.createRandom()
-    const privateKey = nodekey.privateKey.replace(/^0x/, '')
-    const publicKey = nodekey.publicKey.replace(/^0x04/, '')
-    const address = nodekey.address.replace(/^0x/, '').toLowerCase()
 
     return { privateKey, publicKey, address }
   }