+++ src/wallet/model/type/wallet.type.ts
@@ -1,0 +1,9 @@
+export enum WalletType {
+  ETHEREUM = 'ethereum'
+}
+/**
+ * @requires ethereum - [string] ethereum
+*/
+export interface WalletCreateType {
+  type: WalletType
+}